testes contaCorrente

+++ Projeto-Bank-Accenture/src/main/java/com/bankaccenture/Projeto_Bank_Accenture/exception/ContaCorrenteNaoEncontradaException.java
@@ -8,7 +8,7 @@ public class ContaCorrenteNaoEncontradaException extends RuntimeException {
         super(message);
     }
 
-    public ContaCorrenteNaoEncontradaException(Long contaCorrenteId) {
+    public ContaCorrenteNaoEncontradaException(int contaCorrenteId) {
         super("Conta corrente destino não encontrada com ID " + contaCorrenteId);
     }
 }
