+++ Projeto-Bank-Accenture/src/main/java/com/bankaccenture/Projeto_Bank_Accenture/exception/SaldoInsuficienteException.java
@@ -1,5 +1,6 @@
 package com.bankaccenture.Projeto_Bank_Accenture.exception;
 
+import java.math.BigDecimal;
 
 @SuppressWarnings("serial")
 public class SaldoInsuficienteException extends RuntimeException {
@@ -8,7 +9,7 @@ public class SaldoInsuficienteException extends RuntimeException {
         super(message);
     }
 
-    public SaldoInsuficienteException(double valor, double saldoAtual) {
+    public SaldoInsuficienteException(BigDecimal valor, BigDecimal saldoAtual) {
         super("Saldo insuficiente para realizar a operação de valor " + valor + ". Saldo atual: " + saldoAtual);
     }
 }
